requirements.txt
Django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.11
python-decouple==3.8
pytest==8.4.2
pytest-django==4.11.1
pillow==12.0.0
psycopg2-binary==2.9.11
